Build pkgloader as part of "make all"
pkgloader was not built as part of "make all", only indirectly as part
of "make install".  As a result, if one runs "sudo make install", the
"build" subdirectory of the source tree becomes owned by root, which can
lead to annoyance.  Therefore, build pkgloader during "make all",
similar to the skytools package.